rename log/internal/global to log/internal/globallogger
